# Ensure `translation` stage always outputs ISA circuits