--- a/qiskit/transpiler/preset_passmanagers/level2.py
+++ b/qiskit/transpiler/preset_passmanagers/level2.py
@@ -77,15 +77,6 @@ def level_2_pass_manager(pass_manager_config: PassManagerConfig) -> StagedPassMa
         "translation", translation_method, pass_manager_config, optimization_level=2
     )
 
-    if (coupling_map and not coupling_map.is_symmetric) or (
-        target is not None and target.get_non_global_operation_names(strict_direction=True)
-    ):
-        pre_optimization = common.generate_pre_op_passmanager(
-            target, coupling_map, remove_reset_in_zero=False
-        )
-    else:
-        pre_optimization = common.generate_pre_op_passmanager(remove_reset_in_zero=False)
-
     optimization = plugin_manager.get_passmanager_stage(
         "optimization", optimization_method, pass_manager_config, optimization_level=2
     )
@@ -113,7 +104,6 @@ def level_2_pass_manager(pass_manager_config: PassManagerConfig) -> StagedPassMa
         layout=layout,
         routing=routing,
         translation=translation,
-        pre_optimization=pre_optimization,
         optimization=optimization,
         scheduling=sched,
     )