@jakelishman 2025-02-06
I ran the full benchmark suite, and there (as mostly expected) doesn't seem to be much change from this.  There's a small positive improvement in a couple of places of runtime, though I'm not convinced it wasn't just fluctuations on my machine.  There's depth improvements in QUEKO, perhaps because in cases where the optimisation loop is entered twice, there's a bit more information available to it?  Or maybe it's just because there's the knock-on effect of some changes to the randomisation.

```text
| Change   | Before [4d2265f6] <dag-module^2>   | After [92f30cdf] <translation-stage-direction>   |   Ratio | Benchmark (Parameter)                                                    |
|----------|------------------------------------|--------------------------------------------------|---------|--------------------------------------------------------------------------|
| -        | 1.24±0.04ms                        | 1.07±0.04ms                                      |    0.86 | quantum_info.CliffordDecomposeBench.time_decompose('5,10')               |
| -        | 4.53±0.4ms                         | 3.33±0.07ms                                      |    0.73 | quantum_info.SparsePauliOpBench.time_tensor(150, 100)                    |
| -        | 18.0±0.4ms                         | 13.5±0.3ms                                       |    0.75 | queko.QUEKOTranspilerBench.time_transpile_bigd(3, None)                  |
| -        | 81.9±1ms                           | 61.9±1ms                                         |    0.76 | queko.QUEKOTranspilerBench.time_transpile_bntf(0, 'sabre')               |
| -        | 138±6ms                            | 88.2±1ms                                         |    0.64 | queko.QUEKOTranspilerBench.time_transpile_bntf(0, None)                  |
| -        | 114±2ms                            | 99.8±1ms                                         |    0.87 | queko.QUEKOTranspilerBench.time_transpile_bntf(2, 'sabre')               |
| -        | 165±0.7ms                          | 123±2ms                                          |    0.74 | queko.QUEKOTranspilerBench.time_transpile_bntf(3, 'sabre')               |
| -        | 74                                 | 43                                               |    0.58 | queko.QUEKOTranspilerBench.track_depth_bigd_optimal_depth_45(2, 'sabre') |
| -        | 74                                 | 45                                               |    0.61 | queko.QUEKOTranspilerBench.track_depth_bigd_optimal_depth_45(2, None)    |
| -        | 74                                 | 43                                               |    0.58 | queko.QUEKOTranspilerBench.track_depth_bigd_optimal_depth_45(3, 'sabre') |
| -        | 74                                 | 45                                               |    0.61 | queko.QUEKOTranspilerBench.track_depth_bigd_optimal_depth_45(3, None)    |
| -        | 255                                | 216                                              |    0.85 | queko.QUEKOTranspilerBench.track_depth_bntf_optimal_depth_25(2, 'sabre') |
| -        | 45                                 | 30                                               |    0.67 | queko.QUEKOTranspilerBench.track_depth_bntf_optimal_depth_25(2, None)    |
| -        | 45                                 | 30                                               |    0.67 | queko.QUEKOTranspilerBench.track_depth_bntf_optimal_depth_25(3, None)    |
| -        | 28.7±1ms                           | 22.5±0.7ms                                       |    0.79 | random_circuit_hex.BenchRandomCircuitHex.time_ibmq_backend_transpile(10) |
| -        | 39.3±2ms                           | 28.7±1ms                                         |    0.73 | random_circuit_hex.BenchRandomCircuitHex.time_ibmq_backend_transpile(12) |
| -        | 20.4±0.8ms                         | 17.3±0.6ms                                       |    0.85 | random_circuit_hex.BenchRandomCircuitHex.time_ibmq_backend_transpile(8)  |
```